requirements.txt
requests==2.32.5
feedparser==6.0.12
chromadb==1.3.4
openai==2.7.1
python-dotenv==1.2.1
numpy==2.3.4